+++ cli/src/pcluster/config/cluster_config.py
@@ -536,7 +536,7 @@ class PlacementGroup(Resource):
         self.id = Resource.init_param(id)
 
     def _register_validators(self):
-        self._register_validator(PlacementGroupIdValidator, placement_group_id=self.id)
+        self._register_validator(PlacementGroupIdValidator, placement_group=self)
 
 
 class _QueueNetworking(_BaseNetworking):
@@ -1747,9 +1747,7 @@ class SlurmQueue(_CommonQueue):
             self._register_validator(
                 EfaPlacementGroupValidator,
                 efa_enabled=compute_resource.efa.enabled,
-                placement_group_enabled=self.networking.placement_group and self.networking.placement_group.enabled,
-                placement_group_config_implicit=self.networking.placement_group is None
-                or self.networking.placement_group.is_implied("enabled"),
+                placement_group=self.networking.placement_group,
             )
 
     @property